Memorize last commands, suppress sending of patterns
An internal buffer is used to cache the commands that are actually sent
to the deformable mirror.  This avoids the needs to allocate a temporary
array.  The last commands are returned by the `send` method and can be
retrieved with the `lastcommands` method.

Sending of command patterns was not well understood nor well tested has
been suppressed for now.